tech/use gradle kotlin dsl

+++ app/src/main/java/com/prayercompanion/prayercompanionandroid/presentation/di/PresentationModule.kt
@@ -23,7 +23,7 @@ class PresentationModule {
         val activity = context as Activity
         return Doorbell(
             activity,
-            BuildConfig.DOORBELL_ID,
+            BuildConfig.DOORBELL_ID.toLong(),
             BuildConfig.DOORBELL_PRIVATE_KEY,
         )
     }
